refactor: 로또를 관리하는 객체 만들기
